refactor: move utils and types into asset module
