+++ webapp/src/components/ActivityPage/Transaction/TransactionDetail/TransactionDetail.types.ts
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Transaction } from 'decentraland-dapps/dist/modules/transaction/types'
-import { Asset } from '../../../../modules/routing/types'
+import { Asset } from '../../../../modules/asset/types'
 
 export type Props = {
   asset?: Asset | null
