+++ webapp/src/components/AssetAction/AssetAction.types.ts
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Asset } from '../../modules/routing/types'
+import { Asset } from '../../modules/asset/types'
 
 export type Props = {
   asset: Asset
