+++ webapp/src/components/AssetCard/AssetCard.types.ts
@@ -1,4 +1,4 @@
-import { Asset } from '../../modules/routing/types'
+import { Asset } from '../../modules/asset/types'
 import { Order } from '../../modules/order/types'
 
 export type Props = {
