+++ webapp/src/components/AssetImage/AssetImage.tsx
@@ -3,8 +3,8 @@ import { NFTCategory, Rarity } from '@dcl/schemas'
 import { Loader } from 'decentraland-ui'
 import { LazyImage } from 'react-lazy-images'
 
+import { getAssetImage, getAssetName } from '../../modules/asset/utils'
 import { getSelection, getCenter } from '../../modules/nft/estate/utils'
-import { getAssetImage, getAssetName } from '../../modules/nft/utils'
 import { Atlas } from '../Atlas'
 import { Props } from './AssetImage.types'
 import './AssetImage.css'
