+++ webapp/src/components/AssetPage/Actions/Actions.tsx
@@ -4,7 +4,7 @@ import { Network } from '@dcl/schemas'
 import { Modal, Button } from 'decentraland-ui'
 import { T, t } from 'decentraland-dapps/dist/modules/translation/utils'
 
-import { isOwnedBy } from '../../../modules/nft/utils'
+import { isOwnedBy } from '../../../modules/asset/utils'
 import { locations } from '../../../modules/routing/locations'
 import { AssetType } from '../../../modules/asset/types'
 import { VendorFactory } from '../../../modules/vendor'
