+++ webapp/src/components/AssetPage/AssetPage.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { Page } from 'decentraland-ui'
 
-import { Asset } from '../../modules/routing/types'
-import { AssetType } from '../../modules/asset/types'
+import { Asset, AssetType } from '../../modules/asset/types'
 import { Navbar } from '../Navbar'
 import { Footer } from '../Footer'
 import { Navigation } from '../Navigation'
