+++ webapp/src/components/AssetPage/ItemDetail/ItemDetail.tsx
@@ -3,9 +3,9 @@ import { Link } from 'react-router-dom'
 import { Button, Container, Header, Mana, Stats } from 'decentraland-ui'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 
-import { getAssetName, isOwnedBy } from '../../../modules/nft/utils'
 import { formatMANA } from '../../../lib/mana'
 import { locations } from '../../../modules/routing/locations'
+import { getAssetName, isOwnedBy } from '../../../modules/asset/utils'
 import { Row } from '../../Layout/Row'
 import { Column } from '../../Layout/Column'
 import { AssetImage } from '../../AssetImage'
