+++ webapp/src/components/BidPage/BidModal/BidModal.tsx
@@ -8,9 +8,9 @@ import {
   AuthorizationType
 } from 'decentraland-dapps/dist/modules/authorization/types'
 import { hasAuthorization } from 'decentraland-dapps/dist/modules/authorization/utils'
+import { getAssetName, isOwnedBy } from '../../../modules/asset/utils'
 import { toMANA, fromMANA } from '../../../lib/mana'
 import { AssetAction } from '../../AssetAction'
-import { getAssetName, isOwnedBy } from '../../../modules/nft/utils'
 import { getDefaultExpirationDate } from '../../../modules/order/utils'
 import { locations } from '../../../modules/routing/locations'
 import { useFingerprint } from '../../../modules/nft/hooks'
