+++ webapp/src/components/BuyPage/BuyPage.tsx
@@ -8,10 +8,9 @@ import { Footer } from '../Footer'
 import { Wallet as WalletProvider } from '../Wallet'
 import { AssetProviderPage } from '../AssetProviderPage'
 import { NFT } from '../../modules/nft/types'
-import { isOwnedBy } from '../../modules/nft/utils'
+import { isOwnedBy } from '../../modules/asset/utils'
 import { Order } from '../../modules/order/types'
-import { Asset } from '../../modules/routing/types'
-import { AssetType } from '../../modules/asset/types'
+import { Asset, AssetType } from '../../modules/asset/types'
 import { BuyNFTModal } from './BuyNFTModal'
 import { BuyItemModal } from './BuyItemModal'
 import { Props } from './BuyPage.types'
