+++ webapp/src/components/HomePage/Slideshow/Slideshow.types.ts
@@ -1,4 +1,4 @@
-import { Asset } from '../../../modules/routing/types'
+import { Asset } from '../../../modules/asset/types'
 
 export type Props = {
   title: string
