+++ webapp/src/components/TransferPage/TransferPage.tsx
@@ -7,7 +7,7 @@ import { Wallet } from '../Wallet'
 import { AssetProviderPage } from '../AssetProviderPage'
 import { AssetAction } from '../AssetAction'
 import { locations } from '../../modules/routing/locations'
-import { getAssetName, isOwnedBy } from '../../modules/nft/utils'
+import { getAssetName, isOwnedBy } from '../../modules/asset/utils'
 import { AssetType } from '../../modules/asset/types'
 import { Props } from './TransferPage.types'
 import './TransferPage.css'
