+++ webapp/src/modules/asset/types.ts
@@ -1,4 +1,13 @@
+import { Item } from '@dcl/schemas'
+import { NFT } from '../nft/types'
+
 export enum AssetType {
   ITEM = 'item',
   NFT = 'nft'
 }
+
+export type Asset<T extends AssetType = AssetType> = T extends AssetType.NFT
+  ? NFT
+  : T extends AssetType.ITEM
+  ? Item
+  : NFT | Item
