+++ webapp/src/modules/item/actions.spec.ts
@@ -1,7 +1,7 @@
 import { ChainId, Item, Network } from '@dcl/schemas'
 import { buildTransactionPayload } from 'decentraland-dapps/dist/modules/transaction/utils'
 import { formatMANA } from '../../lib/mana'
-import { getAssetName } from '../nft/utils'
+import { getAssetName } from '../asset/utils'
 import { View } from '../ui/types'
 import {
   buyItemFailure,
