+++ webapp/src/modules/routing/selectors.ts
@@ -187,10 +187,9 @@ export const getResultType = createSelector<
   string,
   AssetType
 >(getRouterSearch, getPathName, (search, pathname) => {
-  // TODO: make me pretty
   let results = getURLParam(search, 'assetType') as AssetType
   if (!results) {
-    results = pathname === locations.browse() ? AssetType.ITEM : AssetType.NFT // TODO: i need a isAccountPage or sth helper
+    results = pathname === locations.browse() ? AssetType.ITEM : AssetType.NFT
   }
   return results
 })
